pyproject.toml
[project]
name = "streamlit-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.6",
    "fastembed-gpu>=0.5.0",
    "fastembed>=0.5.0",
    "httpx>=0.28.1",
    "langchain-community>=0.3.13",
    "langchain-huggingface>=0.1.2",
    "langchain-openai>=0.2.14",
    "langchain-qdrant>=0.2.0",
    "markdown>=3.7",
    "matplotlib>=3.10.0",
    "openai>=1.58.1",
    "pip>=24.3.1",
    "pypdf2>=3.0.1",
    "pypdf>=5.1.0",
    "pytest>=8.3.4",
    "qdrant-client>=1.12.2",
    "streamlit>=1.41.1",
    "unstructured>=0.14.8",
    "uvicorn>=0.34.0",
]
